% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\docType{data}
\name{dimensions}
\alias{dimensions}
\title{dimensions}
\format{An object of class \code{character} of length 3.}
\usage{
dimensions
}
\description{
initiate dimension data and create .rda files.
}
\details{
Exported dimension data.frames include: \cr
dim_entities - entity symbol, name, location, region \cr
dim_segments - segment symbol, commodity, group \cr
dim_accounts - account symbol, name, group1, group2 \cr
}
\keyword{dimension}
