% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_data.R
\name{get_base_data}
\alias{get_base_data}
\title{get_base_data function}
\usage{
get_base_data(a = DataArchive, d = Sys.Date())
}
\arguments{
\item{a}{A file path that archives the text files in *YYYYMMDD.TXT naming convention.}

\item{d}{A report control date.}
}
\value{
data A data.frame that contains data from all text files from one control date.
}
\description{
To combine txt files for today, yesterday, last week, last month, last quarter and last year as in report controls table.
}
\details{
@references FileSuffix, DataSeperator, DataHeaders from sysdata.rda.
}
