% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_controls.R
\name{get_controls}
\alias{get_controls}
\title{get_controls function}
\usage{
get_controls(archive = DataArchive)
}
\arguments{
\item{archive}{The file path where data files are archived, default to current directory.}
}
\value{
A data.frame of report run_dates, control periods, control symbols (GOCFxxxxxx).
}
\description{
To generate a list of report controls (report run_date):
today, yesterday, last week, last month, last quarter and last year
It will adjust the dates by checking with archived data files (in *YYYYMMDD.TXT naming convention).
if no data was archived in a date, it will back track until it finds the next data available date.
}
\references{
CtrlPrefix, Filesuffix from sysdata.rda.
}
