% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_date.R
\name{get_report_date}
\alias{get_report_date}
\title{get_report_date function}
\usage{
get_report_date(d = Sys.Date())
}
\arguments{
\item{d}{The report run date or a historical date (report control date), default to current date (Sys.Date).}
}
\value{
A date value
}
\description{
To get a report date for a given run_date, adjusted by weekends and holidays.
}
