% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeper.R
\name{get_timeper}
\alias{get_timeper}
\title{get_timeper function}
\usage{
get_timeper(d = Sys.Date())
}
\arguments{
\item{d}{The run date. it will be converted to position as-of date using get_report_date().}
}
\value{
A data frame of 12 month plus forward time table with these columns:
        MonthId, Month dates, Month symbols, Soybean option months, Canola option months,
        Soymeal option months, Soybean Crush options, Canola Crush options,
        August Old/new Crop, September old/new crop, October old/new crop
}
\description{
To create a twelve month plus forward timeper table for spread position reporting.
}
