% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{main function}
\usage{
main()
}
\description{
This is a sample codes of the main steps to create a data model of fact and dimension tables.
}
\references{
DataSeperator, DataHeaders from sysdata.rda.
}
