% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{sysdata}
\alias{sysdata}
\title{sysdata}
\format{An object of class \code{character} of length 1.}
\usage{
sysdata
}
\description{
Common parameters used in this package
}
\details{
Exported system-wide parameters include: \cr
DataArchive - A file path that source text data files stored. \cr
DataSeperator - The dimiliter of text files, i.e "\{". \cr
DataHeaders - Column names of text files. \cr
CtrlPrefix - The Prefix of dataset control symbol, i.e. GOCF. \cr
FileSuffix  - The suffix as a naming convention of all text files, such as ...20181202.TXT. \cr
klb_to_mt - convert klb to metric tons. \cr
kbu_to_mt - convert k bushels to metric tons. \cr
st_to_mt - convert short tons to metric tons.
}
\keyword{"naming}
\keyword{"package}
\keyword{convention"}
\keyword{parameter",}
